Move TerminalLink
